fix #142: add url 200 tests for dashboard urls
And refactor common bits up into a base test class.
